requirements.txt
numpy==1.18.5
torch==1.7.1
torchvision==0.8.2
pyarrow==3.0.0
tqdm==4.48.2
Pillow==8.3.0
scikit_learn==0.24.2
